Extract request_file method

+++ plexfuse/plex/PlexApi.py
@@ -126,12 +126,8 @@ class PlexApi:
             / self.CACHE_VERSION \
             / item.key.lstrip("/")
 
-    def download_part(self, part, overwrite=False):
-        savepath = self.cache_path(part)
-        if overwrite is False and savepath.exists():
-            return savepath
-
-        url = self.url(part.key)
+    def request_file(self, key: str):
+        url = self.url(key)
         headers = {"X-Plex-Token": self.token}
         response = self.session.get(url, headers=headers, stream=True)
         if response.status_code not in (200, 201, 204):
@@ -139,10 +135,17 @@ class PlexApi:
             message = f"({response.status_code}): {response.url} {errtext}"
             raise RuntimeError(message)
 
-        makedirs(Path(savepath).parent, exist_ok=True)
+        yield from response.iter_content(chunk_size=self.CHUNK_SIZE)
 
+    def download_part(self, part, overwrite=False):
+        savepath = self.cache_path(part)
+        if overwrite is False and savepath.exists():
+            return savepath
+
+        content = self.request_file(part.key)
+        makedirs(Path(savepath).parent, exist_ok=True)
         with open(savepath, "wb") as handle:
-            for chunk in response.iter_content(chunk_size=self.CHUNK_SIZE):
+            for chunk in content:
                 handle.write(chunk)
 
         return savepath
